LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-light-dismiss.tentative-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'document.querySelector('#myElement').shadowRoot.querySelector')
Popover 1 Popover 1 Popover1 anchor (no action) Outside all popovers
Inside popover 1  Popover 2 Inside popover 1 after button
Inside popover 2
Next control after popover1  Popover 3 - button 3
Inside popover 4
Inside popover 3
Inside popover 5  Popover 3 - button 4 - unused
Inside popover 6
Bottom of popover6
Popover 6
Button Inside popover 8 after button
Popover8 anchor (no action)
Popover 1
Anchor  Open Popover 2
Popover 4

Popover 2 (self-anchor-linked)  Open Popover 3  Self-linked invoker
Popover 3  Open Popover 4
Open convoluted popover
Popover 1
Popover 2
Popover 3

Harness Error (FAIL), message = TypeError: null is not an object (evaluating 'document.querySelector('#myElement').shadowRoot.querySelector')

FAIL Clicking outside a popover will dismiss the popover promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Canceling pointer events should not keep clicks from light dismissing popovers promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Clicking inside a popover does not close that popover promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Popovers close on pointerup, not pointerdown promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Synthetic events can't close popovers promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Moving focus outside the popover should not dismiss the popover promise_test: Unhandled rejection with value: object "TypeError: popover1.showPopover is not a function. (In 'popover1.showPopover()', 'popover1.showPopover' is undefined)"
FAIL Clicking inside a child popover shouldn't close either popover promise_test: Unhandled rejection with value: object "TypeError: popover1.showPopover is not a function. (In 'popover1.showPopover()', 'popover1.showPopover' is undefined)"
FAIL Clicking inside a parent popover should close child popover promise_test: Unhandled rejection with value: object "TypeError: popover1.showPopover is not a function. (In 'popover1.showPopover()', 'popover1.showPopover' is undefined)"
FAIL Clicking on invoking element, after using it for activation, shouldn't close its popover promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Clicking on invoking element, after using it for activation, shouldn't close its popover (nested case) promise_test: Unhandled rejection with value: object "TypeError: popover1.showPopover is not a function. (In 'popover1.showPopover()', 'popover1.showPopover' is undefined)"
FAIL Clicking on invoking element, after using it for activation, shouldn't close its popover (nested case, not used for invocation) promise_test: Unhandled rejection with value: object "TypeError: popover1.showPopover is not a function. (In 'popover1.showPopover()', 'popover1.showPopover' is undefined)"
FAIL Clicking on invoking element, even if it wasn't used for activation, shouldn't close its popover promise_test: Unhandled rejection with value: object "TypeError: popover1.showPopover is not a function. (In 'popover1.showPopover()', 'popover1.showPopover' is undefined)"
FAIL Clicking on popovertoggletarget element, even if it wasn't used for activation, should hide it exactly once promise_test: Unhandled rejection with value: object "TypeError: popover1.showPopover is not a function. (In 'popover1.showPopover()', 'popover1.showPopover' is undefined)"
FAIL Clicking on anchor element (that isn't an invoking element) shouldn't prevent its popover from being closed promise_test: Unhandled rejection with value: object "TypeError: popover1.showPopover is not a function. (In 'popover1.showPopover()', 'popover1.showPopover' is undefined)"
FAIL Dragging from an open popover outside an open popover should leave the popover open promise_test: Unhandled rejection with value: object "TypeError: popover1.showPopover is not a function. (In 'popover1.showPopover()', 'popover1.showPopover' is undefined)"
FAIL A popover inside an invoking element doesn't participate in that invoker's ancestor chain promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL An invoking element that was not used to invoke the popover can still be part of the ancestor chain promise_test: Unhandled rejection with value: object "TypeError: popover5.showPopover is not a function. (In 'popover5.showPopover()', 'popover5.showPopover' is undefined)"
FAIL Scrolling within a popover should not close the popover promise_test: Unhandled rejection with value: object "TypeError: popover6.showPopover is not a function. (In 'popover6.showPopover()', 'popover6.showPopover' is undefined)"
FAIL Moving focus back to the anchor element should not dismiss the popover promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL Ensure circular/convoluted ancestral relationships are functional promise_test: Unhandled rejection with value: object "TypeError: convPopover1.showPopover is not a function. (In 'convPopover1.showPopover()', 'convPopover1.showPopover' is undefined)"
FAIL Ensure circular/convoluted ancestral relationships are functional, with a direct showPopover() promise_test: Unhandled rejection with value: object "TypeError: convPopover1.showPopover is not a function. (In 'convPopover1.showPopover()', 'convPopover1.showPopover' is undefined)"
FAIL Hide the target popover during "hide all popovers until" promise_test: Unhandled rejection with value: object "TypeError: p13.showPopover is not a function. (In 'p13.showPopover()', 'p13.showPopover' is undefined)"
